add concurrent batches (serially sending events within a batch) instead of doing one concurrent send per event in the batch.
Before this change we would start sending a batch once one of the triggers (# events/timeout) fired,
but we'd make N calls req.send at once.

This change instead works through an individual batch serially, sending one event then the next.

Concurrency is managed by permitting multiple batches to be processed at the same time, up to maxConcurrentBatches (default 10.)

also before this change our eventQueue could grow without bound if processing events was slow.  This change adds
pendingWorkCapacity (default 10000) which is the max size of the pending event queue before we start dropping events.

This brings the behavior of the JS SDK a little closer in line with the other SDKs.  There is still no blocking on send/response,
as that doesn't really work in the JS context.