fix concurrent batching bugs and add response queue.
found/fixed a bug with the concurrent batching in beta4, and added a test to make sure things work
if we end up requiring more batches than maxConcurrentBatches.

Also add a response queue to the JS SDK (finally.)  The way you access it can be seen in
examples/express/express-honey.js:

```js
    var honey = new libhoney(...);
    const responseCallback = (queue) => {
      var responses = queue.splice(0, queue.length);
      for (var i = 0; i < responses.length; i ++) {
        console.log("response status =", responses[i].status_code);
      }
    };
    honey.once("response", responseCallback);
```

the maximum size of the response queue is supplied in the options passed to libhoney (`maxResponseQueueSize`).
The default value is 1000.  Once there are that many responses, libhoney starts dropping responses (newer
responses are dropped, not older ones.)  calling `.splice(0, queue.length)` is one way to ensure that the
queue is fully drained.  Throttling the response callback might be a good idea so you aren't handling every
response as it comes in.