use /1/batch/<dataset> endpoint (#16)
pretty simple pair of changes masked with a lot of additional churn.

Adds a class called BatchEndpointAggregator to transmission.js that groups events based on apiHost/writekey/datasetname. It does manual JSON encoding to not double encode event post data (which is already JSON.stringifyed before making it to transmission)

The Transmission#_sendBatch method is where the other part of the change is. Once a batch of events is chosen (through timeout or event count triggers), it's aggregated. Then we create a promise to send each batch endpoint aggregated list of events.

The bulk of the test changes are to deal with the fact that with the batch endpoint supported, we get multiple responses at once.