Update test's syntax and improve test readability (#43)
As I've been working on this codebase, I've found the tests a little difficult to work with. This PR changes the assertion syntax to use Jest's standard syntax and also adds some clearer error messages to improve readability.

These changes have come out of my experiment in converting this library into Typescript. Still a work in progress and perhaps might not see the light of day but worth upstreaming these few improvements. Got a few more ready to go soon following this. 😄