add a timeout for an entire batch send/response of 60s (#70)
Most of the fix for https://github.com/honeycombio/beeline-nodejs/issues/201 (will still need to add the config parameter there as well to pass on to libhoney).

Adds a `timeout` config option to the libhoney (and transmission), and respect it when doing a `POST`.

also clean up the superagent-mocker between tests so we can actually have an unmocked test.  superagent-mocker doesn't allow for async mocks 🤯 so we have to run an actual http server as part of this particular test in order to simulate a timeout.