fix: remove yarn from engines (#170)
Fixes #169

Setting the yarn engine to an unacceptable value is a neat trick for
services or internal libraries to use to migrate teams over to a new
engine. That shouldn't be done in public libraries, though! This change
allows downstream consumers of libhoney-js to still be able to use yarn.