Use url-join instead of urljoin
libhoney is pulling in a helper library called urljoin to fuse
URL components together. urljoin expects to be run in a Node
environment, and uses the Node standard library (notably the path module).
Because of this dependency, libhoney fails for us when run under Webpack 5,
which no longer bundles Node libraries in.

This PR swaps urljoin out for an alternative library with equivalent
functionality, called url-join. url-join has no dependencies, and is
newer than urljoin and still maintained.

Because it's new, it uses ES Modules; I needed to configure Jest to
use Babel to transform it.