replace superagent-proxy with direct use of proxy-agent
This is the equivalent behavior of the previous agent configuration. If
a proxy URL is provided in libhoney config, a single ProxyAgent will be
created with that proxy URL set for all connections. No nuanced lookup
of proxy config from the environment based on target URL protocol.

proxy-agent is a hefty import, but we were importing an early edition of
it already via superagent-proxy.