fix: replace superagent-proxy with direct use of proxy-agent (#389)
## Which problem is this PR solving?

- Resolves #380 

## Short description of the changes

This is the equivalent behavior of the previous agent configuration. If
a proxy URL is provided in libhoney config, a single ProxyAgent will be
created with that proxy URL set for all connections. No nuanced lookup
of proxy config from the environment based on target URL protocol.

[proxy-agent is a hefty
import](https://bundlephobia.com/package/proxy-agent@6.3.0), but we were
importing an early edition of it
([v5.0.0](https://bundlephobia.com/package/proxy-agent@5.0.0)) already
via superagent-proxy. 

---------

Co-authored-by: JamieDanielson <jamieedanielson@gmail.com>